Update EditProfile.js
The  "Save profile" section should not be inside      if (profileImage ...). Because, if for some reason the user wants to just update the name but does not want to change his photo, then he wont be able to change his name either. I think the save profile should come right after the if condition.